refactor: 도메인 함수 파일명 및 클래스명 수정
